# docs/swagger.yaml
basePath: /
definitions:
...
  model.ClassroomMemberResponse:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  model.ClassroomResponse:
    properties:
      codes:
        type: string
      cover_img:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_external_invite_enable:
        type: boolean
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      wallpaper_img:
        type: string
    type: object
...
  model.GeneralResponse:
    properties:
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
    type: object
  model.JoinClassroomRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
...
  model.UpdateMemberRoleRequest:
    properties:
      role:
        enum:
        - teacher
        - student
        type: string
    required:
    - role
    type: object
...
host: localhost:9999
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: Baseline project using Go Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Fiber Clean Architecture
  version: 1.0.0
paths:
...
  /api/v1/classrooms:
    get:
      consumes:
      - application/json
      description: list all classrooms.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ClassroomResponse'
                  type: array
              type: object
      security:
      - JWT: []
      summary: list classrooms
      tags:
      - Classroom
    post:
      consumes:
      - multipart/form-data
      description: create a new classroom.
      parameters:
      - description: Classroom Name
        in: formData
        name: name
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Enable External Invite
        in: formData
        name: is_external_invite_enable
        type: boolean
      - description: Cover image file (JPEG, PNG, WebP, max 1MB)
        in: formData
        name: cover_img
        type: file
      - description: Wallpaper image file (JPEG, PNG, WebP, max 1MB)
        in: formData
        name: wallpaper_img
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ClassroomResponse'
              type: object
      security:
      - JWT: []
      summary: create classroom
      tags:
      - Classroom
  /api/v1/classrooms/{id}:
    delete:
      consumes:
      - application/json
      description: delete a classroom by UUID.
      parameters:
      - description: Classroom ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: delete classroom
      tags:
      - Classroom
    get:
      consumes:
      - application/json
      description: get a specific classroom by UUID (user must be a member or admin).
      parameters:
      - description: Classroom ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ClassroomResponse'
              type: object
      security:
      - JWT: []
      summary: get classroom
      tags:
      - Classroom
    put:
      consumes:
      - multipart/form-data
      description: update an existing classroom.
      parameters:
      - description: Classroom ID
        in: path
        name: id
        required: true
        type: string
      - description: Classroom Name
        in: formData
        name: name
        type: string
      - description: Description
        in: formData
        name: description
        type: string
      - description: Enable External Invite
        in: formData
        name: is_external_invite_enable
        type: boolean
      - description: Status
        in: formData
        name: status
        type: string
      - description: Cover image file (JPEG, PNG, WebP, max 1MB)
        in: formData
        name: cover_img
        type: file
      - description: Wallpaper image file (JPEG, PNG, WebP, max 1MB)
        in: formData
        name: wallpaper_img
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ClassroomResponse'
              type: object
      security:
      - JWT: []
      summary: update classroom
      tags:
      - Classroom
  /api/v1/classrooms/{id}/leave:
    post:
      consumes:
      - application/json
      description: leave classroom.
      parameters:
      - description: Classroom ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: leave classroom
      tags:
      - Classroom
  /api/v1/classrooms/{id}/members:
    get:
      consumes:
      - application/json
      description: list classroom members.
      parameters:
      - description: Classroom ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ClassroomMemberResponse'
                  type: array
              type: object
      security:
      - JWT: []
      summary: list classroom members
      tags:
      - Classroom
  /api/v1/classrooms/{id}/members/{userId}:
    delete:
      consumes:
      - application/json
      description: kick classroom member.
      parameters:
      - description: Classroom ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID to kick
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: kick classroom member
      tags:
      - Classroom
    put:
      consumes:
      - application/json
      description: update classroom member role.
      parameters:
      - description: Classroom ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID to update
        in: path
        name: userId
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  type: string
              type: object
      security:
      - JWT: []
      summary: update classroom member role
      tags:
      - Classroom
  /api/v1/classrooms/join:
    post:
      consumes:
      - application/json
      description: join classroom by code.
      parameters:
      - description: Join Classroom Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.JoinClassroomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ClassroomResponse'
              type: object
      security:
      - JWT: []
      summary: join classroom
      tags:
      - Classroom
  /api/v1/classrooms/me:
    get:
      consumes:
      - application/json
      description: list user's classrooms.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.GeneralResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ClassroomResponse'
                  type: array
              type: object
      security:
      - JWT: []
      summary: list my classrooms
      tags:
      - Classroom
...
schemes:
- http
- https
securityDefinitions:
  JWT:
    description: Authorization For JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"